services/analysis_service/requirements.txt
fastapi
gunicorn
uvicorn
google-generativeai
opencv-python 
matplotlib 
pyzbar
